pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saim-Hassan786", email = "saimaiengineer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.73",
    "litellm>=1.61.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
project-1 = "project_1.workflow:main_flow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
